focusadhd-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
SQLAlchemy>=2.0.29
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.9
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sse-starlette>=2.0.0
google-genai[aiohttp]==0.6.0
google-cloud-texttospeech>=2.16.0
google-cloud-aiplatform>=1.48.0
google-cloud-storage>=2.14.0
